Add apt and cloudlab type aggregates.
Add the geni-lib apt and cloudlab aggregates. Stop the madness of
separate if/then for each aggregate type and instead build a single
dictionary of the name mappings. Construct the name mapping in reverse
order so that later name mappings override earlier ones to keep the
result the same as the previos series of if/then statements.