Load RSpecs after all have been fetched
When running in a low memory environment the simultaneous fetching
and loading of RSpecs seems to cause a hang. Change to fetch them
all, then load them all. There isn't much lost parallelism by
loading the RSpecs serially, and the memory consumption does not
interfere with, and hang, the fetching.